requirements.txt
# GPU: NVIDIA GeForce RTX 3050
# CUDA: 11.2
# cuDNN: 8.1.1

cython
h5py
imgaug
IPython
matplotlib
numpy
opencv-python
Pillow
protobuf==3.20
pycocotools
pyyaml==5.4.1
scikit-image==0.16.2
scipy
tensorflow-gpu==2.5.0